test(mint): separate API tests from direct internal ledger method tests
